OJ-1492: move vc types to types folder
