+++ lambdas/issue-credential/src/issue-credential-handler.ts
@@ -17,10 +17,10 @@ import { DynamoDBDocument } from "@aws-sdk/lib-dynamodb";
 import getToyByIdMiddleware from "../../middlewares/toy/get-toy-by-session-id-middleware";
 import setGovUkSigningJourneyIdMiddleware from "../../middlewares/session/set-gov-uk-signing-journey-id-middleware";
 import { VerifiableCredentialBuilder } from "../../verifiable-credential/verifiable-credential-builder";
-import { VC_CONTEXT } from "../../verifiable-credential/types/verifiable-credentials";
+import { VC_CONTEXT } from "../../types/verifiable-credentials";
 import { CommonConfigKey } from "../../types/config-keys";
 import initialiseConfigMiddleware from "../../middlewares/config/initialise-config-middleware";
-import { Subject } from "./subject";
+import { Subject } from "../../types/subject";
 import { ToyItem } from "../../types/toy_item";
 const TOY_CREDENTIAL_ISSUER = "toy_credential_issuer";
 const dynamoDbClient = createClient(AwsClientType.DYNAMO) as DynamoDBDocument;
