+++ lambdas/types/verifiable-credentials.ts
